input page shown
